asdsad

+++ app/models/shop.py
@@ -1,6 +1,7 @@
-from sqlalchemy import String, Integer, Boolean, DateTime, Float
+from sqlalchemy import String, Integer, Boolean, DateTime, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime, timezone
+from decimal import Decimal
 from app.database import Base
 
 
@@ -14,7 +15,7 @@ class Shop(Base):
     description: Mapped[str | None] = mapped_column(String(1000), nullable=True)
     owner_name: Mapped[str] = mapped_column(String(255), nullable=False)
     avatar_url: Mapped[str | None] = mapped_column(String(500), nullable=True)
-    balance: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
+    balance: Mapped[Decimal] = mapped_column(Numeric(10, 2), default=0.0, nullable=False)
     is_approved: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=lambda: datetime.now(timezone.utc), nullable=False)
     updated_at: Mapped[datetime] = mapped_column(
